Move hashing into subpackage
