+++ hash/main.go
@@ -1,4 +1,4 @@
-package stellarbase
+package hash
 
 import (
 	"crypto/sha256"
